2007-11-08  Hiroyuki Ikezoe  <poincare@ikezoe.net>
	* cutter/cut-assertions.c: (cut_assert_equal_int): Cast glong.




git-svn-id: https://cutter.svn.sourceforge.net/svnroot/cutter/trunk@461 10a583ed-1e3d-0410-b285-c80daaaae891

+++ cutter/cut-assertions.h
@@ -110,8 +110,8 @@ void  cut_test_context_register_result      (CutTestContext *context,
         gchar *message;                                                 \
         message = g_strdup_printf(                                      \
             "<" #expected " == " #actual ">\n"                          \
-            "expected: <%d>\n but was: <%d>",                           \
-            (expected), (actual));                                      \
+            "expected: <%ld>\n but was: <%ld>",                         \
+            (glong)(expected), (glong)(actual));                        \
         cut_test_fail(FAILURE, message, ## __VA_ARGS__, NULL);          \
         g_free(message);                                                \
         return;                                                         \
